AWS support

+++ lib/rouster/vagrant.rb
@@ -46,7 +46,13 @@ class Rouster
   # if :sshtunnel is passed to the object during instantiation, the tunnel is created here as well
   def up
     @logger.info('up()')
-    self.vagrant(sprintf('up %s', @name))
+
+    # don't like putting this here, may be refactored
+    if self.is_passthrough? and (self.passthrough[:type].equal?(:aws) or self.passthrough[:type].equal?(:raiden))
+      self.aws_up()
+    else
+      self.vagrant(sprintf('up %s', @name))
+    end
 
     @ssh_info = nil # in case the ssh-info has changed, a la destroy/rebuild
     self.connect_ssh_tunnel() if @sshtunnel
@@ -73,8 +79,15 @@ class Rouster
   # runs `vagrant destroy <name>` from the Vagrantfile path
   def destroy
     @logger.info('destroy()')
+
+    # don't like putting this here, may be refactored
+    if self.is_passthrough? and (self.passthrough[:type].equal?(:aws) or self.passthrough[:type].equal?(:raiden))
+      self.aws_destroy()
+    else
+      self.vagrant(sprintf('destroy -f %s', @name))
+    end
+
     disconnect_ssh_tunnel
-    self.vagrant(sprintf('destroy -f %s', @name))
   end
 
   ##
@@ -94,21 +107,28 @@ class Rouster
       end
     end
 
+    # don't like putting this here, may be refactored
     @logger.info('status()')
-    self.vagrant(sprintf('status %s', @name))
+    if self.is_passthrough? and (self.passthrough[:type].equal?(:aws) or self.passthrough[:type].equal?(:raiden))
+      status = self.aws_status()
+    else
+      self.vagrant(sprintf('status %s', @name))
 
-    # else case here (both for nil/non-matching output) is handled by non-0 exit code
-    output = self.get_output()
-    if output.nil?
-      if self.is_passthrough?()
-        status = 'running'
+      # else case here (both for nil/non-matching output) is handled by non-0 exit code
+      output = self.get_output()
+      if output.nil?
+        if self.is_passthrough?() and self.passthrough[:type].eql?(:local)
+          status = 'running'
+        else
+          status = 'not-created'
+        end
+      elsif output.match(/^#{@name}\s*(.*\s?\w+)\s\((.+)\)$/)
+        # vagrant 1.2+, $1 = status, $2 = provider
+        status = $1
+      elsif output.match(/^#{@name}\s+(.+)$/)
+        # vagrant 1.2-, $1 = status
+        status = $1
       end
-    elsif output.match(/^#{@name}\s*(.*\s?\w+)\s\((.+)\)$/)
-      # vagrant 1.2+, $1 = status, $2 = provider
-      status = $1
-    elsif output.match(/^#{@name}\s+(.+)$/)
-      # vagrant 1.2-, $1 = status
-      status = $1
     end
 
     if @cache_timeout
